# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Capyngen Backend API",
    "description": "API documentation for Capyngen backend (Contact, Blogs, Careers)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local server"
    },
    {
      "url": "https://capyngen-backendv2-1.onrender.com",
      "description": "Render deployed server"
    }
  ],
  "paths": {
    "/api/contact": {
      "post": {
        "summary": "Submit contact form",
        "tags": ["Contact"],
        "requestBody": {
          "description": "Contact form details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "John Doe" },
                  "email": { "type": "string", "example": "john@example.com" },
                  "subject": { "type": "string", "example": "Inquiry" },
                  "message": { "type": "string", "example": "Hi, I want to know more about your services." }
                },
                "required": ["name", "email", "message"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email sent successfully",
            "content": {
              "application/json": {
                "example": { "ok": true, "message": "Sent" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": { "ok": false, "errors": [{ "msg": "Invalid email", "param": "email" }] }
              }
            }
          },
          "500": {
            "description": "Failed to send email",
            "content": {
              "application/json": {
                "example": { "ok": false, "message": "Failed to send email" }
              }
            }
          }
        }
      }
    },
    "/api/blogs": {
      "post": {
        "summary": "Create a new blog post",
        "tags": ["Blogs"],
        "requestBody": {
          "description": "Blog post data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "example": "My First Blog" },
                  "author": { "type": "string", "example": "Admin" },
                  "description": { "type": "string", "example": "A short description of the blog" },
                  "content": { "type": "string", "example": "This is the full blog content." },
                  "image": { "type": "string", "example": "https://example.com/blog.jpg" },
                  "tags": {
                    "type": "array",
                    "items": { "type": "string" },
                    "example": ["React", "Node.js", "MongoDB"]
                  }
                },
                "required": ["title", "author", "description", "content"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Blog saved successfully",
            "content": {
              "application/json": {
                "example": {
                  "ok": true,
                  "message": "Blog saved successfully",
                  "blog": {
                    "_id": "65123abcd456ef7890ghijk",
                    "title": "My First Blog",
                    "author": "Admin",
                    "description": "A short description of the blog",
                    "content": "This is the full blog content.",
                    "image": "https://example.com/blog.jpg",
                    "tags": ["React", "Node.js", "MongoDB"],
                    "createdAt": "2025-09-30T18:30:00.000Z",
                    "updatedAt": "2025-09-30T18:30:00.000Z"
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error" },
          "500": { "description": "Failed to save blog" }
        }
      },
      "get": {
        "summary": "Get all blog posts",
        "tags": ["Blogs"],
        "responses": {
          "200": {
            "description": "List of blogs",
            "content": {
              "application/json": {
                "example": {
                  "ok": true,
                  "blogs": [
                    {
                      "_id": "65123abcd456ef7890ghijk",
                      "title": "My First Blog",
                      "author": "Admin",
                      "description": "A short description of the blog",
                      "content": "This is the full blog content.",
                      "image": "https://example.com/blog.jpg",
                      "tags": ["React", "Node.js", "MongoDB"],
                      "createdAt": "2025-09-30T18:30:00.000Z",
                      "updatedAt": "2025-09-30T18:30:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "500": { "description": "Failed to fetch blogs" }
        }
      }
    },
    "/api/blogs/{id}": {
  "put": {
    "summary": "Update a blog post",
    "tags": ["Blogs"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Blog ID"
      }
    ],
    "requestBody": {
      "description": "Fields to update in the blog",
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "title": { "type": "string", "example": "Updated Blog Title" },
              "author": { "type": "string", "example": "Admin" },
              "description": { "type": "string", "example": "Updated description" },
              "content": { "type": "string", "example": "Updated full content" },
              "image": { "type": "string", "example": "https://example.com/new-image.jpg" },
              "tags": {
                "type": "array",
                "items": { "type": "string" },
                "example": ["Updated", "React"]
              }
            }
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Blog updated successfully",
        "content": {
          "application/json": {
            "example": {
              "ok": true,
              "message": "Blog updated successfully",
              "blog": {
                "_id": "65123abcd456ef7890ghijk",
                "title": "Updated Blog Title",
                "author": "Admin",
                "description": "Updated description",
                "content": "Updated full content",
                "image": "https://example.com/new-image.jpg",
                "tags": ["Updated", "React"],
                "createdAt": "2025-09-30T18:30:00.000Z",
                "updatedAt": "2025-10-01T12:00:00.000Z"
              }
            }
          }
        }
      },
      "404": { "description": "Blog not found" },
      "500": { "description": "Failed to update blog" }
    }
  },
  "delete": {
    "summary": "Delete a blog post",
    "tags": ["Blogs"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Blog ID"
      }
    ],
    "responses": {
      "200": {
        "description": "Blog deleted successfully",
        "content": {
          "application/json": {
            "example": {
              "ok": true,
              "message": "Blog deleted successfully"
            }
          }
        }
      },
      "404": { "description": "Blog not found" },
      "500": { "description": "Failed to delete blog" }
    }
  }
},
    "/api/careers": {
      "post": {
        "summary": "Submit career application",
        "tags": ["Careers"],
        "requestBody": {
          "description": "Career application details",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Jane Doe" },
                  "email": { "type": "string", "example": "jane@example.com" },
                  "position": { "type": "string", "example": "Frontend Developer" },
                  "resume": { "type": "string", "format": "binary" }
                },
                "required": ["name", "email", "position", "resume"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Application submitted",
            "content": {
              "application/json": {
                "example": { "ok": true, "message": "Application received" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "500": { "description": "Failed to submit application" }
        }
      }
    }
  },
  "tags": [
    { "name": "Contact", "description": "Contact form API" },
    { "name": "Blogs", "description": "Blog posting and retrieval API" },
    { "name": "Careers", "description": "Career application API" }
  ]
}